REWORK GAME SERVICE (BACKEND):
Move the gameService to gameplayService in order for better distinction between gameService (get infos about games, eventually managing match making, launch new games etc...) and the gamePlay (the game logic, instantiated by gameService i guess).
