+++ back/src/types/game.ts
@@ -1,5 +1,7 @@
 import { GameState } from '@prisma/client';
-import { Socket } from 'dgram';
+// export * from './gameplay';
+
+export * from './gameplay';
 
 export declare type LadderDisplay = {
 	id: number;
@@ -24,57 +26,3 @@ export declare type GameStats = {
 	winRatio: number;
 	goalRatio: number;
 };
-
-export enum Colour {
-	RED = "RED",
-	WHITE = "WHITE",
-	BLACK = "BLACK",
-};
-
-export enum Move {
-	UP,
-	DOWN,
-	NONE,	
-};
-
-export interface Coord {
-	x: number;
-	y: number;
-};
-
-export interface Rectangle {
-	colour: Colour;
-	width: number;
-	length: number;
-};
-
-export interface Ball {
-	colour: Colour;
-	coord: Coord;
-	vect: Coord; // Contains the ball directional vector
-	radius: number;
-};
-
-export interface Paddle {
-	coord: Coord; // Center of the paddle
-	obj: Rectangle;
-	move: Move;
-};
-
-export interface GameField {
-	field: Rectangle;
-	ball: Ball;
-	paddle1: Paddle;
-	paddle2: Paddle;
-};
-
-export interface Player {
-	userId: number;
-	socketId: Socket; // Handled by Socket IO
-	score: number;
-};
-
-export interface GameResult {
-	winner: Player;
-	looser: Player;
-}
